+++ openeogeotrellis/backend.py
@@ -75,11 +75,8 @@ from openeogeotrellis.job_registry import (
     zk_job_info_to_metadata,
     DoubleJobRegistry,
 )
-from openeogeotrellis.layercatalog import (
-    get_layer_catalog,
-    check_missing_products,
-    GeoPySparkLayerCatalog,
-)
+from openeogeotrellis.layercatalog import (get_layer_catalog, check_missing_products, GeoPySparkLayerCatalog,
+                                           is_layer_too_large, )
 from openeogeotrellis.logs import elasticsearch_logs
 from openeogeotrellis.ml.GeopySparkCatBoostModel import CatBoostClassificationModel
 from openeogeotrellis.sentinel_hub.batchprocessing import (
@@ -896,10 +893,10 @@ class GpsProcessing(ConcreteProcessing):
             if source_id_proc == "load_collection":
                 collection_id = source_id_args[0]
                 metadata = GeopysparkCubeMetadata(catalog.get_collection_metadata(collection_id=collection_id))
+                temporal_extent = constraints.get("temporal_extent")
+                spatial_extent = constraints.get("spatial_extent")
 
                 if metadata.get("_vito", "data_source", "check_missing_products", default=None):
-                    temporal_extent = constraints.get("temporal_extent")
-                    spatial_extent = constraints.get("spatial_extent")
                     properties = constraints.get("properties", {})
                     if temporal_extent is None:
                         yield {"code": "UnlimitedExtent", "message": "No temporal extent given."}
@@ -920,6 +917,25 @@ class GpsProcessing(ConcreteProcessing):
                                 "code": "MissingProduct",
                                 "message": f"Tile {p!r} in collection {collection_id!r} is not available."
                             }
+
+                cell_width = float(metadata.get("cube:dimensions", "x", "step", default = 10.0))
+                cell_height = float(metadata.get("cube:dimensions", "y", "step", default = 10.0))
+                bands = constraints.get("bands", [])
+                geometries = constraints.get("aggregate_spatial", {}).get("geometries")
+                if geometries is None:
+                    geometries = constraints.get("filter_spatial", {}).get("geometries")
+                if is_layer_too_large(
+                    spatial_extent=spatial_extent,
+                    geometries=geometries,
+                    temporal_extent=temporal_extent,
+                    nr_bands=len(bands),
+                    cell_width=cell_width,
+                    cell_height=cell_height
+                ):
+                    yield {
+                        "code": "LayerTooLarge",
+                        "message": "Layer is too large to be processed."
+                    }
 
 
 def get_elastic_job_registry(
@@ -1423,8 +1439,11 @@ class GpsBatchJobs(backend.BatchJobs):
                 jinja_template = Environment(
                     loader=FileSystemLoader(jinja_dir)
                 ).from_string(open(jinja_path).read())
+
+                spark_app_id = k8s_job_name(job_id=job_id, user_id=user_id)
+
                 rendered = jinja_template.render(
-                    job_name=k8s_job_name(job_id=job_id, user_id=user_id),
+                    job_name=spark_app_id,
                     job_specification=job_specification_file,
                     output_dir=output_dir,
                     output_file="out",
@@ -1468,7 +1487,6 @@ class GpsBatchJobs(backend.BatchJobs):
 
                 try:
                     submit_response = api_instance.create_namespaced_custom_object("sparkoperator.k8s.io", "v1beta2", "spark-jobs", "sparkapplications", dict_, pretty=True)
-                    spark_app_id = k8s_job_name(job_id=job_id, user_id=user_id)
                     logger.info(f"mapped job_id {job_id} to application ID {spark_app_id}", extra={'job_id': job_id})
                     dbl_registry.set_application_id(job_id, user_id, spark_app_id)
                     status_response = {}
@@ -2151,7 +2169,7 @@ class GpsBatchJobs(backend.BatchJobs):
                 version = "v1beta2"
                 namespace = "spark-jobs"
                 plural = "sparkapplications"
-                name = k8s_job_name(job_id=job_id, user_id=user_id)
+                name = application_id
                 logger.debug(f"Sending API call to kubernetes to delete job: {name}")
                 delete_response = api_instance.delete_namespaced_custom_object(group, version, namespace, plural, name)
                 logger.debug(